requirements.txt
streamlit
pymupdf
python-docx
scikit-learn
google-genai
